// components/onboarding/types.ts
export type OnboardingData = {
    name?: string
    age?: number
    phone?: string
    city?: string
    state?: string
    language?: string
    employmentType?: string
    monthlyIncome?: number
    employmentTenure?: string
    companyName?: string
    existingEMI?: number
    monthlyExpenses?: number
    savingsRange?: string
    hasCreditHistory?: boolean
    creditScore?: number
    loanPurpose?: string
    loanAmount?: number
    preferredEMI?: number
    tenure?: number
    isJointApplication?: boolean
    coborrowerIncome?: number
    coborrowerRelationship?: string
}